+++ navigation/src/androidMain/kotlin/com/freeletics/khonshu/navigation/NavigationSetup.kt
@@ -1,22 +1,76 @@
-package com.freeletics.khonshu.navigation.internal
+package com.freeletics.khonshu.navigation
 
 import android.app.Activity
+import android.content.Context
+import android.content.ContextWrapper
+import android.os.Parcelable
+import androidx.activity.compose.LocalOnBackPressedDispatcherOwner
+import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.ActivityResultLauncher
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.ProvidableCompositionLocal
+import androidx.compose.runtime.staticCompositionLocalOf
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.LocalLifecycleOwner
 import androidx.core.app.ActivityCompat
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.flowWithLifecycle
-import com.freeletics.khonshu.navigation.ActivityResultRequest
-import com.freeletics.khonshu.navigation.ContractResultOwner
-import com.freeletics.khonshu.navigation.NavEventNavigator
-import com.freeletics.khonshu.navigation.NavigationResultRequest
-import com.freeletics.khonshu.navigation.PermissionsResultRequest
-import com.freeletics.khonshu.navigation.PermissionsResultRequest.PermissionResult
+import com.freeletics.khonshu.navigation.internal.InternalNavigationApi
+import com.freeletics.khonshu.navigation.internal.NavEvent
+import com.freeletics.khonshu.navigation.internal.NavigationExecutor
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import kotlinx.parcelize.Parcelize
+import org.jetbrains.annotations.VisibleForTesting
 
-@InternalNavigationApi
-public suspend fun NavEventNavigator.collectAndHandleNavEvents(
+/**
+ * Sets up the [NavEventNavigator] inside the current composition so that it's events
+ * are handled while the composition is active.
+ */
+@Composable
+public fun NavigationSetup(navigator: NavEventNavigator) {
+    val executor = LocalNavigationExecutor.current
+    val lifecycleOwner = LocalLifecycleOwner.current
+    val context = LocalContext.current
+
+    val activityLaunchers = navigator.activityResultRequests.associateWith {
+        rememberResultLaunchers(it, context)
+    }
+
+    navigator.navigationResultRequests.forEach {
+        LaunchedEffect(executor, it) {
+            executor.collectAndHandleNavigationResults(it)
+        }
+    }
+
+    val backDispatcher = LocalOnBackPressedDispatcherOwner.current!!.onBackPressedDispatcher
+    DisposableEffect(backDispatcher, navigator) {
+        backDispatcher.addCallback(navigator.onBackPressedCallback)
+
+        onDispose {
+            navigator.onBackPressedCallback.remove()
+        }
+    }
+
+    LaunchedEffect(lifecycleOwner, executor, navigator) {
+        navigator.collectAndHandleNavEvents(lifecycleOwner.lifecycle, executor, activityLaunchers)
+    }
+}
+
+@Composable
+private fun <I, O> rememberResultLaunchers(
+    request: ContractResultOwner<I, O, *>,
+    context: Context,
+): ActivityResultLauncher<*> {
+    return rememberLauncherForActivityResult(request.contract) {
+        request.deliverResult(context.findActivity(), it)
+    }
+}
+
+@VisibleForTesting
+internal suspend fun NavEventNavigator.collectAndHandleNavEvents(
     lifecycle: Lifecycle,
     executor: NavigationExecutor,
     activityLaunchers: Map<ContractResultOwner<*, *, *>, ActivityResultLauncher<*>>,
@@ -83,13 +137,14 @@ private fun NavigationExecutor.navigateTo(
     }
 }
 
-@InternalNavigationApi
-public fun <I, O, R> ContractResultOwner<I, O, R>.deliverResult(activity: Activity, result: O) {
+@VisibleForTesting
+internal fun <I, O, R> ContractResultOwner<I, O, R>.deliverResult(activity: Activity, result: O) {
     deliverResult(result) {
         ActivityCompat.shouldShowRequestPermissionRationale(activity, it)
     }
 }
 
+@VisibleForTesting
 @Suppress("UNCHECKED_CAST")
 internal inline fun <I, O, R> ContractResultOwner<I, O, R>.deliverResult(
     result: O,
@@ -100,17 +155,17 @@ internal inline fun <I, O, R> ContractResultOwner<I, O, R>.deliverResult(
         is PermissionsResultRequest -> onResult(
             (result as Map<String, Boolean>).mapValues { (permission, granted) ->
                 if (granted) {
-                    PermissionResult.Granted
+                    PermissionsResultRequest.PermissionResult.Granted
                 } else {
-                    PermissionResult.Denied(shouldShowPermissionRationale(permission))
+                    PermissionsResultRequest.PermissionResult.Denied(shouldShowPermissionRationale(permission))
                 }
             },
         )
     }
 }
 
-@InternalNavigationApi
-public suspend fun <R : Parcelable> NavigationExecutor.collectAndHandleNavigationResults(
+@VisibleForTesting
+internal suspend fun <R : Parcelable> NavigationExecutor.collectAndHandleNavigationResults(
     request: NavigationResultRequest<R>,
 ) {
     val savedStateHandle = savedStateHandleFor(request.key.destinationId)
@@ -126,3 +181,18 @@ public suspend fun <R : Parcelable> NavigationExecutor.collectAndHandleNavigatio
 
 @Parcelize
 private object InitialValue : Parcelable
+
+@InternalNavigationApi
+public val LocalNavigationExecutor: ProvidableCompositionLocal<NavigationExecutor> = staticCompositionLocalOf {
+    throw IllegalStateException("Can't use NavEventNavigationHandler outside of a navigator NavHost")
+}
+
+@InternalNavigationApi
+public fun Context.findActivity(): Activity {
+    var context = this
+    while (context is ContextWrapper) {
+        if (context is Activity) return context
+        context = context.baseContext
+    }
+    throw IllegalStateException("Permissions should be requested in the context of an Activity")
+}